Initial organization feature flag and route
Feature flags can now be toggled in the config/environments.js file.
This allows us to keep the existing redirects to dashboard in production
and staging, but use the new organization panes in development. When
ready, we can just toggle the feature flag in staging and then
production (then remove the feature flag).