Add role permission editing, via Changeset
Changeset is a new utility designed to help us manage arbitrary changes
to the state of a relationship (or even a property, anything really)
that exist only in memory and on the screen and may be disregarded. The
pattern is to only address the relationship via a key made of the
related objects, and change a value in memory. When the user is ready to
save, this list of changed can be iterated and each staged change can be
inspected against its initial value to determine what should be done
(ala create a relationship, delete a relationship, etc).