Add expiry month, expiry year and stripeToken fields to billing-detail.
I had inadvertently removed expiry month and year when extracting the
billing API from auth, so now they're back. In the billing-detail
serializer, I have added a bunch of fields that doesn't need to
be serialized when posting to the billing API.

aptible/dashboard.aptible.com#382 depends on this commit.