Add support for clustered databases
The UI should behave a little differently for databases that support
replication and databases that support clustering. For example,
while it's fine to daisy-chain replicated databases, it doesn't really
make sense to do so in the clustered case. Furthermore, the terminology
is different.

Note that since the backend doesn't really understand the difference
between clustering and replication (all it cares about is what existing
database the new one should initialize from), it's worth trying to keep
the relationships in good order on the frontend (i.e. avoid
daisy-chaining in the clustered scenario), but it's not stricly-speaking
required for things to work.