Drag-n-drop .crt, bundled .crt, and .key files to create certificates
Closes https://github.com/aptible/dashboard.aptible.com/issues/532

I did a bit of hallway user testing on this (thanks @shahkader &
@wcpines ) in hopes this is an improved user experience. You can drop
multiple files on the textarea and that makes it tricky to order
bundled cert chains, but with
https://github.com/aptible/api.aptible.com/pull/327, the ordering will
be taken care of on the server and the user will not need to worry
about it.

Currently we don’t notify users of ordering problems and they end up
seeing them in a browser
(https://support.aptible.com/topics/paas/how-to-order-certs/).

We may be better served by a traditional file upload button. Will watch
for feedback.