Shared FileTextarea component
Appends a textarea with the contents of whatever file was dropped on
it. See https://github.com/aptible/dashboard.aptible.com/pull/543 for
details.

Note to self — lost a lot of time moving this from diesel, shared
components’ templates need to go in `app/templates/components/` not in
the component directory like they do in diesel. :(