C3Chart: Unload existing data when updating
Since we're introducing metrics where a simple toggle introduces data
with a new ID, we need to make sure old data is properly collected to
avoid leaving unwanted data on the graph.

This also fixes an invalid check for `data` being empty (which didn't
cause any issues, but is probably better fixed).