Fix test crashes due to c3chart cleanup
The c3 chart component sometimes fails in test when the chart is
destroyed while a transition is running (the transition runs
asynchronously and attempts to access attributes that are removed in
the `destroy` method).

This is an issue in test, although it's not really an issue live
(because the chart isn't usually dismissed so fast, and even if it is,
then it's not going to actually crash the app).

This works around the issue by delaying the destroy operation, which
isn't a perfect solution, but is the only one that only affects tests
without adding too much complexity (see comments for more details).