Fix non-resolving promises in billing overview
This is a commit that adjusts some issues with #580. In some cases the
promises were never fully resolved in the UI, which led to some
calculations being incorrect. I refactored the promise resolving so
that it all happens within the route instead of trying to rely on
computed properties to resolve the promises.

The rest is pretty much simply refactoring so that the affected
components are easier to test, as well as changing the rendering of
vhost rows to only render if its status is provisioned, and only
rendering disk rows if the disk has a non empty size.