Require elevation on user settings route
This wraps the user settings route below a require-elevation route. The
requires-elevation route enforces that the current token has the
`elevated` scope, and if not, redirects to the `elevate` route.

The `elevate` route requests that the user input their credentials again
(it uses the same form and logic as the login route), and calls the Auth
API to generate a new short-lived (30 minutes) elevated token using
those. If that succeeds, it logs the user out and logs them back in with
the new token.

The elevated token is never persisted to cookies (the request to create
it is made with `withCredentials:  false`), and is discarded as soon as
the user transitions to a route that does not require elevation. In
practice, this means that as soon as the user is no longer seeing the
settings page, then the elevated token is discarded.

The Auth API enforces that updates to a user's email, password, or 2FA
settings require an elevated token.