Avoid 2-way binding on the user settings page
The user settings has 3 forms that modify the same record, which makes
2-way binding a bad fit. For example, saving an updated password can:

- Persist the new OTP configuration on the user (the backend now rejects
  that change without a valid OTP token if OTP is enabled, but that
  means your password update is rejected because you don't have a valid
  OTP token even though you haven't enabled it and shouldn't be asked
  for it!).
- Persist an updated email. The backend might even reject the change if
  the email is invalid...

Disabling 2-way binding is the simpler way to resolve those issues. It's
also a reasonable change to make since the tests were added while we had
2-way binding, and didn't change (except for the addition of new tests).

---

This commit also fixes an issue when leaving the settings page without
an OTP configuration, which would attempt to clear a field that did not
exist.