Ensure OTP token is discarded after input
If a user enables OTP, then disables it, then starts enabling it again,
then the token they entered the first time around will still be shown.
This update ensures the OTP token is discarded whenever the perform a
change to the user's OTP status.

The same applies to showing backup codes, so this also resets this
parameter.