Fix log drain count header
We show host:port for one log drain in the stack header, but if the log
drain is a log tail, that breaks down (because it does not have those).

This updates the logic to prefer showing a log drain that's not a log
tail, and shows nothing if the log drain is a log tail.