signIn helper: don't access session directly
In practice, this function has been a recurring problem for me; since it
only accepts a few very specific parameters that it loads into the user,
token, and session, it's never quite up to date with the rest of the
code, and tests end up failing because the stub isn't stubbing things
properly.

Getting rid of the stub entirely, and relying on `/current_token`
instead seems like a better option.