Simplify impersonation workflow
This creates the user token and then reloads the UI in order to use it.
It's arguably a little safer (since we never access a user's data before
destroying our admin token), but it's also a lot more convenient, since
it avoids putting the UI into an awkward state if the token exchange
fails.

To do so, we need to add a `noClearCookies` flag when closing a session,
which allows us to clear our admin token, all the while keeping our user
token in store.