Expose a noClearCookies flag when closing session
This also refactors the session argument to be an options hash rather
than just a token, since we're starting to add multiple arguments.

The use case for the noClearCookies flag is to allow impersonation to
create a user token, persist it to cookies, destroy the admin token
(without touching cookies), and then reload the UI to log back in as the
user.