bs-XXXX: `options.html` should not be a function
Bootstrap explicitly accepts functions for the `title`, `content` and
`placement` attributes. However, it most notably does not accept a
function for the `html` attribute, which controls whether the input
provided to e.g. `tooltip()` or `popover()` should be output as HTML.

This updates bootstrap-component-options to actually set a boolean. It
also exposes the option in `bs-tooltip`, `bs-popover`, and
`more-info-icon`.