Make popovers disapper when user navigates away
Anchoring the popovers to the application makes them stick around when
the user navigates away, and anchoring them to the current element makes
them adopt a broken CSS, so this anchors them to the backup list, which
resolved both issues.

Unfortunately, it also results in duplication of the `backup-list`
constant name, and generally speaking the `backup-restore-link` aware of
the page that's embedding it...

Unfortunately, I haven't found a great way to pass the value for
bs-container down to the popover.