Fix CSP violations on sandboxes / staging
We're hardcoding a lit of allowed connect-src in our CSP, but it's
incomplete and doesn't work for sandboxes. This updates our CSP
connect-src rules to be dynamically constructed using the endpoint URLs
we're already dynamically generating.

(It also fixes a typo in `metricsBaseuri`, which should be
`metricsBaseUri` for consistency with everything else)