Add logout page under settings
The logout page is a good place to prompt the user to revoke other
tokens, if only so that they know where to find it later. Using an
intermediary logout page also ensures users cannot be tricked into
logging out simply by visiting a link.

However, this does change the UX slightly, and might be unexpected for
existing users. Prior to rolling this out, we should set up an
Announcement via Intercom to notify users of the upcoming change.

Sicne the logout page is now integrated into the settings portal
(probably the only part of the UI that is fully functional for all
users; including those that aren't members of any organization), we also
need to slighly tweak the verification-code-reset component in order to
avoid flashing the yellow "unverified" warning when logging the user
out, simply by not showing it if there is no current user (which is fine
anyway; it wouldn't be functional in that case).

This also adds test coverage for logout, as well as some basic tests for
the verification-code-reset component.