acmeDomain -> isAcme + userDomain + acmeStatus
This refactors acmeDomain into an explicit isAcme flag and a userDomain
field as well as an acmeStatus field that represents whether further
actions are needed.

This will make it easier to prompt users for a domain in cases that
aren't ACME, like provisioning an OAuth authentication proxy. It'll also
allow e.g. deploying a transitional cert to an ACME endpoint.