Catch 404s on reload
This adds another check on top of #117 - this checks explicitly for
err.status == 404, otherwise log drains would not be removed after
deprovisioning them.

There's a minor WTF in this, which I don't really understand. There's
a heisenbug of sorts - now, unless there's a console.log or a
window.alert before the status check, the model will otherwise
immediately remove itself from the UI without regarding the status
that gets returned from the API.