Check for currentState before reloading
This builds upon #154. During a deprovisioning, the model will
periodically poll itself to see if it has been deleted from the api,
where it will then remove itself from the local model storage.

This checks the currentState of the model to ensure that it doesn't
reload if it is in either state == 'root.empty' or in state
'root.deleted.uncommitted'.