Reload app on invitation accept
We cache organization permissions when loading the authorization
service, but when accepting an invitation, the user's permissions are
going to change.

We'd need to either invalidate the cached permissions, or invalidate
everything. This patch does the latter in the interest of un-breaking
accepting invitations, but the former might be a much cleaner approach.