Certificates: Create -> Upload
The copy for uploading a certificate suggests you can create a new
certificate via Aptible, which could be misunderstood as signing a CSR
and returning a Certificate.

While the copy is technically accurate (we create a certificate record
in API), it's a little confusing, since you need to already have created
the certificate somewhere else to be able to use in on Aptible.