New Endpoint: use service process type, not handle
The handle for a service isn't ideal to let users differentiate between
services, mainly because:

- It includes a lot of redundant or useless information (namely, it
  includes the app handle and a random string).
- It is truncated to 32 characters, so it might not even include the
  relevant information (the process type).

This is presumably why we don't use it anywhere in the UI (at least that
I know of), except on the New Endpoint page. This can be a problem,
because it means users sometimes can't tell which service is which.

This patch fixes the issue by instead using the process type and
appending the command (so that it's shown if there is room), which
should make it easier for users to interact with this.

I've also added tests to ensure this selector actually works (which was
a little more involved than I anticipated, since the mocks here were
being created in two different places).